// futurenhs.app/components/forms/CheckBox/interfaces.ts
export interface Props {
    input: {
        name: string
        value: string
        onChange: any
    }
    meta: {
        error: string
        submitError: string
        touched: boolean
    }
    initialError?: string
    text: {
        label: string
        hint?: string
        error?: string
    }
    validators?: Array<any>
    className?: string
}